+++ packages/bezier/CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/tkofh/squiggles/compare/@squiggles/bezier@0.1.0...@squiggles/bezier@0.1.1) (2022-05-08)
+## [0.1.1](https://github.com/tkofh/curvy/compare/@curvy/bezier@0.1.0...@curvy/bezier@0.1.1) (2022-05-08)
 
-**Note:** Version bump only for package @squiggles/bezier
+**Note:** Version bump only for package @curvy/bezier
 
 
 
@@ -16,4 +16,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **bezier:** readme ([04b3046](https://github.com/tkofh/squiggles/commit/04b3046d9ab0d18e9baada53e944677e7bd21ff9))
+* **bezier:** readme ([04b3046](https://github.com/tkofh/curvy/commit/04b3046d9ab0d18e9baada53e944677e7bd21ff9))