+++ packages/dx/CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/tkofh/squiggles/compare/@squiggles/dx@0.1.0...@squiggles/dx@0.1.1) (2022-05-08)
+## [0.1.1](https://github.com/tkofh/curvy/compare/@curvy/dx@0.1.0...@curvy/dx@0.1.1) (2022-05-08)
 
-**Note:** Version bump only for package @squiggles/dx
+**Note:** Version bump only for package @curvy/dx
 
 
 
@@ -16,4 +16,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ### Bug Fixes
 
-* **dx:** use correct project name in package.json ([c9a15b5](https://github.com/tkofh/squiggles/commit/c9a15b589045d3947682774f7f1b746e6dac6298))
+* **dx:** use correct project name in package.json ([c9a15b5](https://github.com/tkofh/curvy/commit/c9a15b589045d3947682774f7f1b746e6dac6298))