+++ packages/math/CHANGELOG.md
@@ -3,9 +3,9 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.1.1](https://github.com/tkofh/squiggles/compare/@squiggles/math@0.1.0...@squiggles/math@0.1.1) (2022-05-08)
+## [0.1.1](https://github.com/tkofh/curvy/compare/@curvy/math@0.1.0...@curvy/math@0.1.1) (2022-05-08)
 
-**Note:** Version bump only for package @squiggles/math
+**Note:** Version bump only for package @curvy/math
 
 
 
@@ -13,4 +13,4 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 # 0.1.0 (2022-05-08)
 
-**Note:** Version bump only for package @squiggles/math
+**Note:** Version bump only for package @curvy/math